implemented user authentication

+++ app/Http/Controllers/ListingController.php
@@ -46,6 +46,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->user()->id;
+
         Listing::create($formFields);
 
         // Session::flash('message', 'Your listing has been added!');
@@ -63,6 +65,12 @@ class ListingController extends Controller
 
     public function update(Request $request, Listing $listing) {
         // dd($listing->id);
+
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, "Unauthorized");
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -86,7 +94,17 @@ class ListingController extends Controller
 
     //delete listing
     public function delete(Listing $listing) {
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, "Unauthorized");
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    //Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
